pyproject.toml
[tool.poetry]
name = "yourml"
version = "0.0.0"
description = "Description of this project"
authors = ["YourName"]
readme = "README.md"
repository = "https://github.com/username/reponame"

[tool.poetry.dependencies]
python = "^3.7"
numpy = ">=1.17"
pytorch-lightning = "^1.4.9"
fsspec = {version = ">=2021.5", extras = ["http", "s3"]}
boto3 = "^1.16.18"
tqdm = "^4.0.0"
omegaconf = "^2.1.1"
speechcorpusy = {git = "https://github.com/tarepan/speechcorpusy.git", rev = "main"}
speechdatasety = {git = "https://github.com/tarepan/speechdatasety.git", rev = "main"}
## Dependencies for Audio
# pyloudnorm = "^0.1.0"
# librosa = "^0.8.0"
## Poetry is not good for version control of PyTorch (it has many many variants for a version, so poetry become confused)
## torch = "1.11.0"
## torchaudio = "*"

[tool.poetry.dev-dependencies]
types-tqdm = "^4.64.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
disable = [
  # This is opinionated
  "line-too-long",
]